+++ packages/react/src/components/test-utils.ts
@@ -132,11 +132,15 @@ export const polyfillPointerEvent = () => {
  * with ease, without the need of accessing private SlashID emitter property.
  */
 export class MockSlashID extends SlashID {
-  private observers: Map<keyof PublicReadEvents, GenericHandler[]>;
+  private _observers: Map<keyof PublicReadEvents, GenericHandler[]> = new Map()
+
+  private get observers() {
+    if (!this._observers) this._observers = new Map()
+    return this._observers
+  }
 
   constructor(opts?: SlashIDOptions) {
     super(opts);
-    this.observers = new Map();
   }
 
   public subscribe<Key extends keyof PublicReadEvents>(