+++ apps/react-vite/CHANGELOG.md
@@ -1,5 +1,14 @@
 # react-vite
 
+## 0.0.46
+
+### Patch Changes
+
+- Updated dependencies [5de98fa]
+- Updated dependencies [c1b0305]
+  - @slashid/react@1.16.0
+  - demo-form@0.0.39
+
 ## 0.0.45
 
 ### Patch Changes